# api/openapi.yaml
openapi: 3.1.0
info:
  title: ReFlow API
  version: 0.1.0
  description: Versioned API shared by ReFlow desktop, PWA, and future mobile clients.
  license:
    name: GPL-3.0-only
servers:
  - url: http://127.0.0.1:7381
paths:
...
  /api/v1/research/papers:
    get:
      operationId: listResearchPapers
      tags: [research]
      parameters:
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum: [research, submitted, published]
      responses:
        "200":
          description: Research workspace papers of the requested kind.
        "400":
          $ref: "#/components/responses/Problem"
    post:
      operationId: createResearchPaper
      tags: [research]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind]
              properties:
                kind:
                  type: string
                  enum: [research, submitted, published]
                title: { type: string, maxLength: 500 }
                authors:
                  type: array
                  items: { type: string, maxLength: 500 }
      responses:
        "201":
          description: Paper created.
        "400":
          $ref: "#/components/responses/Problem"
  /api/v1/research/papers/reorder:
    post:
      operationId: reorderResearchPapers
      tags: [research]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, paper_ids]
              properties:
                kind:
                  type: string
                  enum: [research, submitted, published]
                paper_ids:
                  type: array
                  maxItems: 500
                  items: { type: string }
      responses:
        "204":
          description: Order updated.
        "400":
          $ref: "#/components/responses/Problem"
  /api/v1/research/papers/{paperId}:
    parameters:
      - name: paperId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getResearchPaper
      tags: [research]
      responses:
        "200":
          description: >
            A single research workspace paper. Its tag_ids are every label
            assigned to it, in assignment order, and never null.
        "404":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateResearchPaper
      tags: [research]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: >
                Partial update. Text fields accept at most 500 characters
                (notes and abstract at most 2000), submission_count must not
                be negative, and a paper wears at most 12 tags; violations
                return 400 with code invalid_research_paper.
              properties:
                title: { type: string, maxLength: 500 }
                authors: { type: array, items: { type: string, maxLength: 500 } }
                keywords: { type: array, items: { type: string, maxLength: 500 } }
                file_path: { type: string, maxLength: 500 }
                next_action: { type: string, maxLength: 500 }
                notes: { type: string, maxLength: 2000 }
                research_area: { type: string, maxLength: 500 }
                status: { type: string, maxLength: 500 }
                tag_ids:
                  type: array
                  maxItems: 12
                  items: { type: string, maxLength: 500 }
                  description: >
                    Ids of tags from /api/v1/research/tags. A paper wears them
                    all at once, and the array order is the order the user
                    assigned them in, not the palette order. Repeats collapse,
                    an empty array clears every label, and omitting the field
                    leaves the assignment untouched. Unknown ids are accepted
                    and simply render as nothing until the palette is reloaded.
                target_journal: { type: string, maxLength: 500 }
                stages:
                  type: array
                  items:
                    type: object
                    properties:
                      name: { type: string, maxLength: 500 }
                      done: { type: boolean }
                      children: { type: array, items: { type: object } }
                current_journal: { type: string, maxLength: 500 }
                submission_date: { type: string, maxLength: 500 }
                manuscript_id: { type: string, maxLength: 500 }
                submission_count: { type: integer, minimum: 0 }
                target_level: { type: string, maxLength: 500 }
                editor: { type: string, maxLength: 500 }
                deadline: { type: string, maxLength: 500 }
                history:
                  type: array
                  items:
                    type: object
                    properties:
                      journal: { type: string, maxLength: 500 }
                      date: { type: string, maxLength: 500 }
                      status: { type: string, maxLength: 500 }
                abstract: { type: string, maxLength: 2000 }
                journal: { type: string, maxLength: 500 }
                language: { type: string, maxLength: 500 }
                year: { type: string, maxLength: 500 }
                volume: { type: string, maxLength: 500 }
                issue: { type: string, maxLength: 500 }
                pages: { type: string, maxLength: 500 }
                doi: { type: string, maxLength: 500 }
                citations: { type: [integer, "null"] }
                citation_source: { type: string, maxLength: 500 }
      responses:
        "200":
          description: Updated paper.
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteResearchPaper
      tags: [research]
      responses:
        "204":
          description: Paper deleted.
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/research/papers/{paperId}/move:
    parameters:
      - name: paperId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: moveResearchPaper
      tags: [research]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind]
              properties:
                kind:
                  type: string
                  enum: [research, submitted, published]
      responses:
        "200":
          description: Paper moved to the new kind. Stages and submission history are preserved.
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /api/v1/research/papers/{paperId}/citation:
    parameters:
      - name: paperId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: fetchResearchCitation
      tags: [research]
      description: >
        Refreshes the citation count from Crossref. Returns 422 with code
        citation_no_doi when the paper has no usable DOI, or doi_not_found
        when Crossref has no work for the DOI; 502 crossref_unavailable is
        retryable.
      responses:
        "200":
          description: Citation count refreshed from Crossref.
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "502":
          $ref: "#/components/responses/Problem"
  /api/v1/research/tags:
    get:
      operationId: listResearchTags
      tags: [research]
      description: >
        The profile's tag palette in priority order. The three legacy priority
        labels are seeded by migration 0016 at positions 0, 1 and 2. Drag the
        palette into a new order with POST /api/v1/research/tags/reorder.
      responses:
        "200":
          description: Tags ordered by position.
    post:
      operationId: createResearchTag
      tags: [research]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 40
                  description: Trimmed; a name already used by the profile returns 409.
      responses:
        "201":
          description: Tag appended to the end of the palette.
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /api/v1/research/tags/reorder:
    post:
      operationId: reorderResearchTags
      tags: [research]
      description: >
        Rewrites the palette to the given order so the settings list can be
        dragged. Ids outside the profile are ignored, and a paper's own tag_ids
        keep their assignment order across a palette reorder.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag_ids]
              properties:
                tag_ids:
                  type: array
                  maxItems: 500
                  items: { type: string }
      responses:
        "204":
          description: Palette order updated.
        "400":
          $ref: "#/components/responses/Problem"
  /api/v1/research/tags/{tagId}:
    parameters:
      - name: tagId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: renameResearchTag
      tags: [research]
      description: >
        Renames a tag. Papers reference tags by id, so their tag_ids are
        unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 40
      responses:
        "200":
          description: Renamed tag.
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteResearchTag
      tags: [research]
      description: >
        Deletes a tag and lifts just that label from every paper wearing it;
        the other labels a paper wears keep their order.
      responses:
        "204":
          description: Tag deleted and removed from every paper.
        "404":
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      bearerFormat: ReFlowDeviceToken
...
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        request_id:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
security:
  - deviceToken: []